Minor improvements

+++ main_2d.py
@@ -567,7 +567,7 @@ def iterative_search(m_clean: np.ndarray, m_limits: list[list[int]], par: Parame
 
     # Initialize an array to store labels for each window.
     num_windows = int(m_clean.shape[1] / tau_w)
-    all_the_labels = np.zeros((m_clean.shape[0], num_windows))
+    all_the_labels = np.zeros((m_clean.shape[0], num_windows)).astype(int)
 
     states_list = []
     m_copy = m_clean
